+++ README.md
@@ -1,0 +1,17 @@
+## Rapoholic
+App which connects people who want to sell rap CDs
+
+## How to run
+- Clone repository
+- Go to project folder. In terminal run command *gradle clean build*
+- Go to build folder and run in terminal *java -jar rapoholic.jar*
+
+or
+
+- Clone repository
+- Go to project folder. In terminal run command *gradle bootrun*
+
+## How to access h2 database console
+- In browser go to *localhost:8080/h2-console*
+- As *JDBC URL* set *jdbc:h2:mem:testdb*
+- Log in as *sa*
