+++ docs/useful-commands.txt
@@ -4,3 +4,9 @@ gradle clean build (Budowanie aplikacji)
 
 Git:
 
+How to install Spring boot application with .jar packaging on external server
+https://docs.spring.io/spring-boot/docs/current/reference/html/deployment-install.html
+
+Debugging spring boot application with gradle
+gradle bootRun --debug-jvm
+
