--- gradle.properties
@@ -1,4 +1,0 @@
-springBootStarterVersion=2.0.0.RELEASE
-hibernateCoreVersion=5.2.15.Final
-liquibaseCoreVersion=3.5.5
-liquibaseGradlePlugin=1.1.1
