--- src/main/java/Rapoholic.java
@@ -1,6 +1,0 @@
-
-public class Rapoholic {
-    public boolean someLibraryMethod() {
-        return true;
-    }
-}
