--- src/main/java/model/Plyty.java
@@ -1,45 +1,0 @@
-package model;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name="plyty")
-public class Plyty {
-
-    @Id
-    @Column(name="id")
-    private int id;
-
-    @Column(name="title")
-    private String title;
-
-    @Column(name="release_year")
-    private int releaseYear;
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public void setReleaseYear(int releaseYear) {
-        this.releaseYear = releaseYear;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public int getReleaseYear() {
-        return releaseYear;
-    }
-}
