+++ src/main/java/pl/rapoholic/RapoholicApplication.java
@@ -1,0 +1,12 @@
+package pl.rapoholic;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RapoholicApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RapoholicApplication.class, args);
+	}
+}
