+++ src/main/java/pl/rapoholic/entity/Plyty.java
@@ -1,0 +1,46 @@
+package pl.rapoholic.entity;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "plyty")
+public class Plyty {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "title")
+    private String title;
+
+    @Column(name = "release_year")
+    private int releaseYear;
+
+    public Plyty() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public int getReleaseYear() {
+        return releaseYear;
+    }
+
+    public void setReleaseYear(int releaseYear) {
+        this.releaseYear = releaseYear;
+    }
+}
