+++ src/test/java/pl/rapoholic/RapoholicApplicationTests.java
@@ -1,0 +1,16 @@
+package pl.rapoholic;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class RapoholicApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
